# maquinaria_laravel/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Maquinaria",
        "description": "Documentación de la API de Mantenimientos y Maquinarias",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/detalles-solicitud": {
            "get": {
                "tags": [
                    "DetalleSolicitud"
                ],
                "summary": "Listar todos los detalles de solicitud",
                "operationId": "e289435470a3baebced6e9de71d5b72f",
                "responses": {
                    "200": {
                        "description": "Lista de detalles de solicitud",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/DetalleSolicitud"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "DetalleSolicitud"
                ],
                "summary": "Crear un detalle de solicitud",
                "operationId": "9d7895f748f9fbe40d3401800fc62902",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "solicitud_id",
                                    "mantenimiento_id",
                                    "cantidad_maquinas",
                                    "Url_foto"
                                ],
                                "properties": {
                                    "solicitud_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "mantenimiento_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "cantidad_maquinas": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "Url_foto": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Detalle de solicitud creado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DetalleSolicitud"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/detalles-solicitud/{id}": {
            "get": {
                "tags": [
                    "DetalleSolicitud"
                ],
                "summary": "Obtener un detalle de solicitud por ID",
                "operationId": "0366f8d3ef137566df78637e7df001e9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del detalle de solicitud",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalle encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DetalleSolicitud"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Detalle no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "DetalleSolicitud"
                ],
                "summary": "Actualizar un detalle de solicitud",
                "operationId": "a95d158d8406e14406f88dfd26dc1368",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "cantidad_maquinas": {
                                        "type": "integer",
                                        "example": 4
                                    },
                                    "mantenimiento_id": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Detalle actualizado"
                    },
                    "404": {
                        "description": "Detalle no encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "DetalleSolicitud"
                ],
                "summary": "Eliminar un detalle de solicitud",
                "operationId": "568cdadc58935874a44bc92efea169e1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalle eliminado"
                    },
                    "404": {
                        "description": "Detalle no encontrado"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "DetalleSolicitud": {
                "title": "Detalle de Solicitud",
                "description": "Modelo que representa un detalle dentro de una solicitud de mantenimiento",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "solicitud_id": {
                        "description": "ID de la solicitud relacionada",
                        "type": "integer",
                        "example": 10
                    },
                    "mantenimiento_id": {
                        "description": "ID del mantenimiento relacionado",
                        "type": "integer",
                        "example": 5
                    },
                    "cantidad_maquinas": {
                        "description": "Número de máquinas solicitadas",
                        "type": "integer",
                        "example": 3
                    },
                    "costo_total": {
                        "description": "Costo total del detalle",
                        "type": "number",
                        "format": "float",
                        "example": 1500.5
                    },
                    "Url_foto": {
                        "description": "Ruta de la imagen subida",
                        "type": "string",
                        "example": "uploads/imagenes/maquina1.jpg"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}